requirements.txt
click==6.7
colorama==0.3.9
GitPython==2.1.5